refactor(frontend): split the credit-note section into a ledger and a form
Aikido read InvoiceCreditNotes as one 166-line component doing credit
calculations, form validation, action handling, permission checks and
ledger rendering.

CreditNoteLedger owns the list and the void control, CreditNoteIssueForm
owns the amount and reason fields and their validation, and what is left
is 106 lines that compose the two and hold the credited total.

Full frontend suite still green at 978 suites and 12144 tests, all three
components at 100% statements, branches, functions and lines, and the 16
Storybook stories render clean in both themes.

Cognitive complexity checked locally with eslint-plugin-sonarjs run
directly, since it is wired into no eslint config: clean across every
file this branch touches.
